requirements.txt
asgiref==3.6.0
beautifulsoup4==4.12.2
cssutils==2.7.1
Django==4.2
flake8==6.1.0
future==0.18.3
mccabe==0.7.0
mysqlclient==2.1.1
Pillow==9.5.0
pycodestyle==2.11.0
pyflakes==3.1.0
pynliner==0.8.0
python-decouple==3.8
soupsieve==2.4.1
sqlparse==0.4.4
